requirements.txt
PyQt5>=5.15.0
PyQt5-sip>=12.8.0
pyserial>=3.5
lupa>=2.4
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
numpy>=1.24.0 